+++ src/api/run.js
@@ -49,6 +49,7 @@ export default function (store = {}, option = {}) {
     middlewares, plugins, isStrict, isDebug, errorHandler, isHot,
     autoCreateDispatcher, reducer, bindCtxToMethod,
     computedCompare, watchCompare, watchImmediate,
+    generatorReducer,
   } = option;
 
   if (reducer) {
@@ -66,7 +67,7 @@ export default function (store = {}, option = {}) {
   // merge startupOption
   Object.assign(startupOption, {
     middlewares, plugins, isStrict, isDebug, errorHandler, isHot, autoCreateDispatcher, bindCtxToMethod,
-    computedCompare, watchCompare, watchImmediate,
+    computedCompare, watchCompare, watchImmediate, generatorReducer,
   });
 
   startup(startupOption);